fix: move the category deletion to the gc
